requirements.txt
fpdf==1.7.2
pdf2image==1.17.0
Pillow==9.5.0
PyPDF2==3.0.1
